% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_label.R, R/ai_topic.R
\name{ai_label}
\alias{ai_label}
\title{Labels documents in a corpus according to a content analysis scheme}
\usage{
ai_label(.data, chat_fn, ..., label, verbose = TRUE)

ai_label(.data, chat_fn, ..., label, verbose = TRUE)
}
\arguments{
\item{.data}{a character or [quanteda::corpus] object containing the
documents to be summarized}

\item{chat_fn}{function; a chat function from \pkg{ellmer}}

\item{...}{additional arguments passed to `chat_fn`}

\item{label}{the defined content analysis scheme for the labels - be as specific as possible}

\item{verbose}{logical; output a progress indicator if `TRUE`}

\item{model}{a llm model object}
}
\value{
character; the response from the LLM with a length equal to the
  number of input documents, with the elements named with the input element
  names

character; the response from the LLM with a length equal to the
  number of input documents, with the elements named with the input element
  names
}
\description{
Use the llm model to analyse a set of documents. The text-based response is
added as an additional column to the input data.

Use the llm model to analyse a set of documents. The text-based response is
added as an additional column to the input data.
}
\examples{
\dontrun{
library(quanteda)
label = "Label the following document based on how much it aligns with 
  the political left, center, or right. The political left is defined as groups 
  which advocate for social equality, government intervention in the economy, 
  and progressive policies. The political center typically supports a balance 
  between progressive and conservative views, favoring moderate policies       
  and compromise. The political right generally advocates for individualism, 
  free-market capitalism, and traditional values."
label1 <- ai_label(data_corpus_inaugural[1:2], chat_fn = chat_ollama, model = "llama3.2", label = label)
label2 <- ai_label(data_char_ukimmig2010, chat_fn = chat_openai,
               api_args = list(temperature = 0, seed = 42), label = label)
}
\dontrun{
library(quanteda)
label = "Label the following document based on how much it aligns with 
  the political left, center, or right. The political left is defined as groups 
  which advocate for social equality, government intervention in the economy, 
  and progressive policies. The political center typically supports a balance 
  between progressive and conservative views, favoring moderate policies       
  and compromise. The political right generally advocates for individualism, 
  free-market capitalism, and traditional values."
label1 <- ai_label(data_corpus_inaugural[1:2], chat_fn = chat_ollama, model = "llama3.2", label = label)
label2 <- ai_label(data_char_ukimmig2010, chat_fn = chat_openai,
               api_args = list(temperature = 0, seed = 42), label = label)
}
}
