% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_quant.R
\name{ai_quant}
\alias{ai_quant}
\title{Score documents using a llm model
Use the llm model to score a set of documents.
The numerical scores for each text are added as an additional docvar/column to the input data}
\usage{
ai_quant(.data, texts, model, concept)
}
\arguments{
\item{.data}{a \pkg{quanteda} corpus or data frame containing the documents to be summarized}

\item{texts}{the name of the docvar/column in the corpus/data frame containing the documents to be summarized}

\item{model}{a llm model object}

\item{concept}{the concept based on which the llm should score how much the documents align to it - be as specific as possible}
}
\value{
the input \pkg{quanteda} corpus or data frame with the document scores stored as an additional docvar/column
}
\description{
Score documents using a llm model
Use the llm model to score a set of documents.
The numerical scores for each text are added as an additional docvar/column to the input data
}
\examples{
\dontrun{
concept = "The political left defined as groups which advocate for social equality,
           government intervention in the economy, and progressive policies."
ai_quanties <- ai_quant(corpus_split, "splits", "llama3.2", concept)
}
}
