% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_sum.R
\name{ai_sum}
\alias{ai_sum}
\title{Summarize documents using a llm model
Use the llm model to summarize a set of documents.
The response is added as an additional docvar to the input data}
\usage{
ai_sum(.data, texts, model, length, strict = FALSE)
}
\arguments{
\item{.data}{a data frame containing the documents to be summarized}

\item{texts}{the name of the column in the data frame containing the documents to be summarized}

\item{model}{a llm model object}

\item{length}{the number of words to include in the summary}

\item{strict}{a logical indicating whether the summary should be exactly `length` words, default is `FALSE`}
}
\value{
the input data frame with the document summaries stored as an additional column
}
\description{
Summarize documents using a llm model
Use the llm model to summarize a set of documents.
The response is added as an additional docvar to the input data
}
\examples{
\dontrun{
#ai_summary <- ai_sum(corpus_split, "splits", "llama3.2", 100)
}

}
