% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{dialogic_analysis_module}
\alias{dialogic_analysis_module}
\title{Dialogic analysis module factory}
\usage{
dialogic_analysis_module()
}
\value{
An llm_module for dialogic analysis
}
\description{
Dialogic analysis module factory
}
