% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{factuality_module}
\alias{factuality_module}
\title{Factuality checking module factory}
\usage{
factuality_module()
}
\value{
An llm_module for factuality checking
}
\description{
Factuality checking module factory
}
