% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{named_entity_module}
\alias{named_entity_module}
\title{Named entity recognition module factory}
\usage{
named_entity_module(
  named_entity_categories = c("people", "parties", "organizations")
)
}
\arguments{
\item{named_entity_categories}{Character vector of categories (default c("people", "parties", "organizations"))}
}
\value{
An llm_module for named entity recognition
}
\description{
Named entity recognition module factory
}
