% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{narrative_module}
\alias{narrative_module}
\title{Narrative analysis module factory}
\usage{
narrative_module()
}
\value{
An llm_module for narrative analysis
}
\description{
Narrative analysis module factory
}
