% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{position_taking_module}
\alias{position_taking_module}
\title{Position taking module factory}
\usage{
position_taking_module(stance_labels = c("support", "oppose", "neutral"))
}
\arguments{
\item{stance_labels}{Character vector of stance labels (default c("support", "oppose", "neutral"))}
}
\value{
An llm_module for position taking
}
\description{
Position taking module factory
}
