% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.llm-package.R
\docType{package}
\name{quanteda.llm-package}
\alias{quanteda.llm}
\alias{quanteda.llm-package}
\title{quanteda.llm: LLM support for quanteda corpus objects}
\description{
Provides support for linking quanteda corpus objects to structured, flexible analysis by LLMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quanteda/quanteda.llm}
  \item Report bugs at \url{https://github.com/quanteda/quanteda.llm/issues}
}

}
\author{
\strong{Maintainer}: Seraphine F. Maerz \email{seraphine.maerz@unimelb.edu.au} (\href{https://orcid.org/0000-0002-7173-9617}{ORCID})

Authors:
\itemize{
  \item Kenneth Benoit \email{kbenoit@smu.edu.sg} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID})
}

}
\keyword{A}
\keyword{LLMs.}
\keyword{\pkg{quanteda}}
\keyword{extension}
\keyword{for}
\keyword{internal}
\keyword{package}
\keyword{with}
\keyword{working}
