% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{tradeoff_module}
\alias{tradeoff_module}
\title{Trade-off identification module factory}
\usage{
tradeoff_module()
}
\value{
An llm_module for trade-off identification
}
\description{
Trade-off identification module factory
}
