% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{translation_module}
\alias{translation_module}
\title{Translation module factory}
\usage{
translation_module(target_language = "English")
}
\arguments{
\item{target_language}{Character string for target language (default "English")}
}
\value{
An llm_module for translation
}
\description{
Translation module factory
}
