% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_validate.R
\name{ai_validate}
\alias{ai_validate}
\title{Starts an interactive app to manually validate the
documents in a corpus which were summarized, labelled, or scored by an LLM.}
\usage{
ai_validate(text, llm_output, ..., verbose = TRUE)
}
\arguments{
\item{llm_output}{a character string; the name of the LLM output column which contains
the summaries, labels, or scores to be validated}

\item{verbose}{logical; output a progress indicator if `TRUE`}

\item{.data}{a character or [quanteda::corpus] object containing the
documents to be manually validated}
}
\value{
character; the response from the manual validation with a length equal to the
  number of input documents, with the elements named with the input element
  names
}
\description{
The comments from manual validation for each text are added as an additional docvar to the input data
with all texts not yet validated as "N/A".
}
\examples{
\dontrun{
library(quanteda)
summ1 <- ai_summarize(data_char_ukimmig2010, chat_fn = chat_ollama, model = "llama3.2")
summ2 <- ai_summarize(data_corpus_inaugural[1:2], chat_fn = chat_openai,
                api_args = list(temperature = 0, seed = 42))
validate1 <- ai_validate(data_char_ukimmig2010, llm_output = summ1, verbose = TRUE)
validate2 <- ai_validate(data_corpus_inaugural[1:2], llm_output = summ2, verbose = TRUE)
}
}
